refactor: ResponseEntity 체이닝 메서드 개행 방식 변경
